image_processor/requirements.txt
pika==1.3.1
requests==2.31.0
Pillow==10.1.0
